requirements.txt
# Web Framework & Async
fastapi==0.135.1
uvicorn[standard]==0.42.0
python-multipart==0.0.22

# Agentic Orchestration
langgraph==1.1.2
langchain==1.2.12
langchain-google-vertexai==3.2.2

# Data & Validation
pydantic==2.12.5
pydantic-settings==2.13.1

# Vector DB & RAG
qdrant-client==1.17.1
pymupdf==1.27.2

# Communication & API
grpcio==1.78.0
grpcio-tools==1.78.0
google-cloud-aiplatform==1.141.0

# Observability & Monitoring
arize-phoenix==13.15.0
openinference-instrumentation-langchain==0.1.61

# Infrastructure
python-dotenv==1.2.2